moves components to profiles folder
